feat(ux): flatten `wash ctl start` into `wash start`
