+++ crates/wash-lib/src/cli/mod.rs
@@ -38,6 +38,7 @@ use crate::{
 pub mod claims;
 pub mod inspect;
 pub mod link;
+pub mod start;
 
 /// Used for displaying human-readable output vs JSON format
 #[derive(Debug, Copy, Clone, Eq, Serialize, Deserialize, PartialEq)]
