+++ crates/wash-lib/src/lib.rs
@@ -28,9 +28,11 @@ pub mod parser;
 #[cfg(feature = "start")]
 pub mod start;
 
+pub mod common;
 pub mod config;
 pub mod context;
 pub mod drain;
 pub mod id;
 pub mod keys;
 pub mod registry;
+pub mod wait;
