+++ src/common/mod.rs
@@ -1,1 +1,2 @@
 pub(crate) mod link_cmd;
+pub(crate) mod start_cmd;
